lingering rename in tests
DiabetesQuery -> DmQuery
