+++ gluco-check-core/test/fakes/objects/fakeDmQuery.ts
